V4L2Camera: Check that the camera accept the frame size as well not only the pixel format
USB UVC cameras may support different frame size sets for
different pixel formats. So do not limitate to check that
the current format is supported and then set the frame size
but check that this frame size _with_ this pixel format works.

Real case is a camera that accepts only QCIF in YUYV but can do
VGA in MJPEG

Signed-off-by: Alberto Panizzo <alberto@amarulasolutions.com>